[Refactor] Merge raycluster_gcs_ft_test.go and raycluster_gcsft_test.go
